+++ docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Click the plus icon under an equipped item to search for and transfer items in that slot from anywhere in your inventory.
 * Import a CSV file of items with tags and notes to bulk update the tags/notes for all of those items.
+* CSV - Wrap ID in quotes such that its value is not rounded.
 
 # 5.8.3 (2019-01-02)