+++ docs/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Next
 
+# 5.9.0 (2019-01-06)
+
 * Click the plus icon under an equipped item to search for and transfer items in that slot from anywhere in your inventory.
 * Import a CSV file of items with tags and notes to bulk update the tags/notes for all of those items.
 * CSV - Wrap ID in quotes such that its value is not rounded.